aws_data_collectors/requirements.txt
# AWS 데이터 수집기 의존성 패키지

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# HTTP 요청
requests>=2.28.0

# 데이터 처리
pandas>=1.5.0
numpy>=1.21.0

# 설정 파일
PyYAML>=6.0

# 파일 처리
pathlib2>=2.3.0

# 로깅
loguru>=0.6.0

# 타입 힌트
typing-extensions>=4.0.0

# 웹 스크래핑
scrapy>=2.8.0
selenium>=4.0.0

# 이미지 처리
Pillow>=9.0.0

# 테스트 (개발용)
pytest>=7.0.0
pytest-cov>=3.0.0

# 코드 품질 (개발용)
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
